# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CurrentUserRole = typeof CurrentUserRole[keyof typeof CurrentUserRole];


export const CurrentUserRole = {
  OWNER: 'OWNER',
  MANAGER: 'MANAGER',
  STAFF: 'STAFF',
} as const;

export interface Business {
  id: number;
  name: string;
  slug: string;
  branchName: string;
}

export interface CurrentUser {
  id: string;
  name: string;
  email: string;
  role: CurrentUserRole;
  business: Business;
}

export interface DashboardStats {
  active: number;
  today: number;
  completed: number;
  averageResponseSeconds: number;
}

export type ServiceRequestPriority = typeof ServiceRequestPriority[keyof typeof ServiceRequestPriority];


export const ServiceRequestPriority = {
  NORMAL: 'NORMAL',
  HIGH: 'HIGH',
  URGENT: 'URGENT',
} as const;

export type ServiceRequestStatus = typeof ServiceRequestStatus[keyof typeof ServiceRequestStatus];


export const ServiceRequestStatus = {
  REQUESTED: 'REQUESTED',
  ACCEPTED: 'ACCEPTED',
  IN_PROGRESS: 'IN_PROGRESS',
  COMPLETED: 'COMPLETED',
  DECLINED: 'DECLINED',
  CANCELLED: 'CANCELLED',
} as const;

export interface ServiceRequest {
  id: number;
  trackingToken: string;
  tableName: string;
  tableArea: string;
  serviceName: string;
  serviceIcon: string;
  priority: ServiceRequestPriority;
  status: ServiceRequestStatus;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  assignedTo?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export type StaffSummaryStatus = typeof StaffSummaryStatus[keyof typeof StaffSummaryStatus];


export const StaffSummaryStatus = {
  AVAILABLE: 'AVAILABLE',
  BUSY: 'BUSY',
  OFFLINE: 'OFFLINE',
} as const;

export interface StaffSummary {
  id: number;
  name: string;
  department: string;
  status: StaffSummaryStatus;
}

export interface Dashboard {
  business: Business;
  stats: DashboardStats;
  requests: ServiceRequest[];
  staff: StaffSummary[];
}

export interface Table {
  id: number;
  code: string;
  name: string;
  area: string;
  active: boolean;
  guestPath: string;
  requestCount: number;
}

export interface TableInput {
  /**
     * @minLength 1
     * @maxLength 40
     */
  name: string;
  /**
     * @minLength 1
     * @maxLength 60
     */
  area: string;
}

export type ServicePriority = typeof ServicePriority[keyof typeof ServicePriority];


export const ServicePriority = {
  NORMAL: 'NORMAL',
  HIGH: 'HIGH',
  URGENT: 'URGENT',
} as const;

export interface Service {
  id: number;
  name: string;
  icon: string;
  department: string;
  priority: ServicePriority;
  active: boolean;
}

export type RequestStatusInputStatus = typeof RequestStatusInputStatus[keyof typeof RequestStatusInputStatus];


export const RequestStatusInputStatus = {
  ACCEPTED: 'ACCEPTED',
  IN_PROGRESS: 'IN_PROGRESS',
  COMPLETED: 'COMPLETED',
  DECLINED: 'DECLINED',
  CANCELLED: 'CANCELLED',
} as const;

export interface RequestStatusInput {
  status: RequestStatusInputStatus;
}

export interface GuestTable {
  businessName: string;
  businessSlug: string;
  branchName: string;
  tableName: string;
  tableCode: string;
  services: Service[];
}

export interface GuestRequestInput {
  serviceId: number;
  /** @maxLength 240 */
  note?: string;
}

export type ListRequestsParams = {
status?: ListRequestsStatus;
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};

export type ListRequestsStatus = typeof ListRequestsStatus[keyof typeof ListRequestsStatus];


export const ListRequestsStatus = {
  REQUESTED: 'REQUESTED',
  ACCEPTED: 'ACCEPTED',
  IN_PROGRESS: 'IN_PROGRESS',
  COMPLETED: 'COMPLETED',
  DECLINED: 'DECLINED',
  CANCELLED: 'CANCELLED',
} as const;