# lib/api-zod/src/generated/types/business.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Business {
  id: number;
  name: string;
  slug: string;
  branchName: string;
}