# lib/api-zod/src/generated/types/currentUser.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Business } from './business';
import type { CurrentUserRole } from './currentUserRole';

export interface CurrentUser {
  id: string;
  name: string;
  email: string;
  role: CurrentUserRole;
  business: Business;
}