# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Business } from './business';
import type { DashboardStats } from './dashboardStats';
import type { ServiceRequest } from './serviceRequest';
import type { StaffSummary } from './staffSummary';

export interface Dashboard {
  business: Business;
  stats: DashboardStats;
  requests: ServiceRequest[];
  staff: StaffSummary[];
}