# lib/api-zod/src/generated/types/guestTable.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Service } from './service';

export interface GuestTable {
  businessName: string;
  businessSlug: string;
  branchName: string;
  tableName: string;
  tableCode: string;
  services: Service[];
}