# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ServicePriority } from './servicePriority';

export interface Service {
  id: number;
  name: string;
  icon: string;
  department: string;
  priority: ServicePriority;
  active: boolean;
}