# lib/api-zod/src/generated/types/serviceRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ServiceRequestPriority } from './serviceRequestPriority';
import type { ServiceRequestStatus } from './serviceRequestStatus';

export interface ServiceRequest {
  id: number;
  trackingToken: string;
  tableName: string;
  tableArea: string;
  serviceName: string;
  serviceIcon: string;
  priority: ServiceRequestPriority;
  status: ServiceRequestStatus;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  assignedTo?: string | null;
  createdAt: Date;
  updatedAt?: Date;
}