# lib/api-zod/src/generated/types/staffSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StaffSummaryStatus } from './staffSummaryStatus';

export interface StaffSummary {
  id: number;
  name: string;
  department: string;
  status: StaffSummaryStatus;
}