# lib/api-zod/src/generated/types/table.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Table {
  id: number;
  code: string;
  name: string;
  area: string;
  active: boolean;
  guestPath: string;
  requestCount: number;
}